vix: the exec backend is injected, not defaulted
`vix-core` declared the ExecBackend seam and then quietly filled it in:
`PrimitiveServices::exec_backend()` fell back to `HostExecBackend` for anyone
who installed nothing. So the design read as injected while the MACHINE was
deciding that vix spawns host processes — invisibly, for every caller.

The default is gone and `HostExecBackend` moved to vixen-runtime, beside the
rest of the assembly (`default_primitive_dispatcher`, `install_builtins`).
`vix-core` now ships no exec backend, the same as it ships no primitives, no
prelude, and no host types. Whoever wants a process boundary says so:
`harness_services()` says so for the test harness, and `vx` says so for the CLI
— which is the point, since swapping in a sandboxing backend is now a one-line
change at the CLI rather than a change to the machine.

Nothing keys on the backend, so this re-keys no demand; a snapshot without one
fails loudly at the first exec (`EffectCtx::exec_backend` already said so — it
just could not be reached while the default existed).

`archive_directory` stays in vix-core with a note: the exec PRIMITIVE calls it
and vixen-primitives cannot reach vixen-runtime. Moving capture behind the
trait is the right end state and changes the trait's shape, so it is not
bundled here.
